+++ tests/testthat/test-get_aux.R
@@ -1,25 +1,26 @@
-test_that("get_countries() works", {
-  res <- get_countries()
-  res2 <- get_aux("countries")
+#Disable caching
+Sys.setenv("PIPR_DISABLE_CACHING" = "TRUE")
+#To enable running tests which uses real API uncomment the below line
+#Sys.setenv("NOT_CRAN" = "true")
+
+## Test core functions ----
+test_that("get_aux returns available tables when no argument is specified", {
+  skip_if_offline()
+  skip_on_cran()
+
+  res <- suppressMessages(get_aux())
   expect_true(tibble::is_tibble(res))
-  expect_identical(res, res2)
+  expect_equal(names(res), "tables")
 })
 
-test_that("get_regions() works", {
-  res <- get_regions()
-  res2 <- get_aux("regions")
-  expect_true(tibble::is_tibble(res))
-  expect_identical(res, res2)
-})
 
-test_that("get_aux() works", {
-
+test_that("get_aux() works when calling specific tables", {
+  skip_if_offline()
+  skip_on_cran()
   # Return tibble as default
-  res <- get_aux()
+  res <- suppressMessages(get_aux())
   expect_true(tibble::is_tibble(res))
   res <- get_aux("gdp")
-  expect_true(tibble::is_tibble(res))
-  res <- get_aux("countries", version = NULL, api_version = "v1", format = "rds", server = NULL)
   expect_true(tibble::is_tibble(res))
 
   # Return custom response list if simplify FALSE
@@ -41,8 +42,8 @@ test_that("get_aux() works", {
   # expect_error(get_aux("tmp"))
   # expect_true(is.list(get_aux("tmp", simplify = FALSE)))
   skip_if(Sys.getenv("PIPR_RUN_LOCAL_TESTS") != "TRUE")
-  expect_error(get_aux("tmp", server = "qa"))
-  expect_true(is.list(get_aux("tmp", simplify = FALSE, server = "qa")))
+  expect_error(get_aux("wrong-table-name", server = "qa"))
+  expect_true(is.list(get_aux("wrong-table-name", simplify = FALSE, server = "qa")))
 
   # Check all tables
   skip("survey_metadata gives a 500 error. Need to add functionality for list data")
@@ -53,6 +54,8 @@ test_that("get_aux() works", {
 })
 
 test_that("User agent works", {
+  skip_if_offline()
+  skip_on_cran()
   # res <- get_aux(simplify = FALSE)
   # tmp <- res$response$request$options$useragent
   # expect_identical(tmp, pipr_user_agent)
@@ -61,3 +64,102 @@ test_that("User agent works", {
   expect_identical(tmp, pipr_user_agent)
 })
 
+## Test helper functions ----
+test_that("get_countries() works", {
+  skip_if_offline()
+  skip_on_cran()
+  res <- get_countries()
+  res2 <- get_aux("countries")
+  expect_true(tibble::is_tibble(res))
+  expect_identical(res, res2)
+})
+
+test_that("get_regions() works", {
+  skip_if_offline()
+  skip_on_cran()
+  res <- get_regions()
+  res2 <- get_aux("regions")
+  expect_true(tibble::is_tibble(res))
+  expect_identical(res, res2)
+})
+
+
+# test_that("get_countries() with mocking works", {
+#   mockery::stub(get_aux, "httr::GET", function(...) {
+#     readRDS(test_path("testdata", "response-country.RDS"))
+#   })
+#
+#   mockery::stub(get_countries, "get_aux", function(...) {
+#     readRDS(test_path("testdata", "response-country.RDS"))
+#   })
+#
+#   res1 <- get_aux("countries")
+#   res2 <- parse_response(get_countries(), TRUE)
+#
+#   expect_true(tibble::is_tibble(res1))
+#   expect_true(tibble::is_tibble(res2))
+#
+#   expect_equal(dim(res1), dim(res2))
+#   expect_identical(res1, res2)
+# })
+#
+#
+# test_that("get_regions() with mocking works", {
+#   mockery::stub(get_aux, "httr::GET", function(...) {
+#     readRDS(test_path("testdata", "response-regions.RDS"))
+#   })
+#
+#   mockery::stub(get_regions, "get_aux", function(...) {
+#     readRDS(test_path("testdata", "response-regions.RDS"))
+#   })
+#
+#   res1 <- get_aux("regions")
+#   res2 <- parse_response(get_regions(), TRUE)
+#
+#   expect_true(tibble::is_tibble(res1))
+#   expect_true(tibble::is_tibble(res2))
+#
+#   expect_equal(dim(res1), dim(res2))
+#   expect_identical(res1, res2)
+# })
+#
+#
+# test_that("get_cpi() with mocking works", {
+#   mockery::stub(get_aux, "httr::GET", function(...) {
+#     readRDS(test_path("testdata", "response-cpi.RDS"))
+#   })
+#
+#   mockery::stub(get_cpi, "get_aux", function(...) {
+#     readRDS(test_path("testdata", "response-cpi.RDS"))
+#   })
+#
+#   res1 <- get_aux("cpi")
+#   res2 <- parse_response(get_cpi(), TRUE)
+#
+#   expect_true(tibble::is_tibble(res1))
+#   expect_true(tibble::is_tibble(res2))
+#
+#   expect_equal(dim(res1), dim(res2))
+#   expect_identical(res1, res2)
+# })
+#
+# test_that("get_dictionary() with mocking works", {
+#   #Waiting for this PR to be merged https://github.com/worldbank/pipr/pull/43
+#   #so that get_dictionary() works
+#   mockery::stub(get_aux, "httr::GET", function(...) {
+#     readRDS(test_path("testdata", "response-dictionary.RDS"))
+#   })
+#
+#   mockery::stub(get_dictionary, "get_aux", function(...) {
+#     readRDS(test_path("testdata", "response-dictionary.RDS"))
+#   })
+#
+#   res1 <- get_aux("dictionary")
+#   res2 <- parse_response(get_dictionary(), TRUE)
+#
+#   expect_true(tibble::is_tibble(res1))
+#   expect_true(tibble::is_tibble(res2))
+#
+#   expect_equal(dim(res1), dim(res2))
+#   expect_identical(res1, res2)
+# })